requirements.txt
requests~=2.22.0
python-dateutil~=2.8.2
pytz~=2022.7.1
pandas~=1.5.3
Flask~=2.2.3
APScheduler~=3.10.1
tradingview-ta